+++ src/main/java/com/example/secondapriltry/SecondapriltryApplication.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.secondapriltry;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
