+++ src/main/java/mapper/ClientRowMapper.java
@@ -1,6 +1,6 @@
-package com.example.restserviceproject.mapper;
+package mapper;
 
-import com.example.restserviceproject.entity.Client;
+import entity.Client;
 import org.springframework.jdbc.core.RowMapper;
 
 import java.sql.ResultSet;
@@ -11,8 +11,8 @@ public class ClientRowMapper implements RowMapper <Client> {
     public Client mapRow(ResultSet resultSet, int i) throws SQLException {
         Client client = new Client();
         client.setId(resultSet.getInt("id"));
-        client.setFirstName(resultSet.getString("firstname"));
-        client.setLastName(resultSet.getString("lastname"));
+        client.setFirstName(resultSet.getString("firsname"));
+        client.setLastName(resultSet.getString("lastName"));
         client.setSurName(resultSet.getString("surname"));
         return client;
     }
