+++ src/main/java/service/ClientService.java
@@ -1,9 +1,10 @@
-package com.example.restserviceproject.dao;
+package service;
 
-import com.example.restserviceproject.entity.Client;
+import entity.Client;
+
 import java.util.List;
 
-public interface ClientDao {
+public interface ClientService {
     List<Client> findAll();
     void insertClient(Client client);
     void updateClient(Client client);
