+++ src/test/java/com/example/secondapriltry/SecondapriltryApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example;
+package com.example.secondapriltry;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
